Allow adding a post-consent CSP for security (#11)
Previously CSP was used purely to block certain elements that required 
consent for marketing cookies and there was no CSP in place after 
consent had been given. This change makes it possible to supply a 
second CSP that is in place after consent so there are always 
limits in place.